High: Fencing: Standardize on /one/ method for clients to request members be fenced
Allow the uname to be provided if the caller knows it, otherwise default
to the corosync node id.

Remove logging from stonith_api_time, let the caller do some if they
want to.

Rename the functions to be less corosync specific (none have appeared in
a release so now is the time).