Fix: services : ppoll if SIGNALFD is not available
SIGNALFD is a pure linux functionality.
ppoll is supported in all OS where pacemaker compiles.
The functionality of the patch should be the same as with the usage of SIGNALFD.
If the child process returns the signal is catched and the loop will be left.
If the process hangs the loop is finished after the the timeout.