Fix: configure.ac : no extra handling for Solaris …
Solaris/Illumos have in /bin/sh an old version of sh which is not compatible with the version of linux.
There is a GNU compatible sh available and must be used when compiling.
With
export CONFIG_SHELL=/usr/gnu/bin/sh
the auto* facility is informed to use this version of sh.
Unfortunately the configure created with configure.ac has as shebang line
#! /bin/sh
although CONFIG_SHELL is defined and exported.
config.status has the expected shebang line of
#! /usr/gnu/bin/sh
I think this is a bug in autoconf.
If configure is started with
/usr/gnu/bin/sh configure ...
the function cc_supports_flag of configure is able to recognize of an option is supported.
Starting with /bin/sh as defined in the shebang line the result of cc_supports_flag is always true because local is not available.